refactor(audio): split recorder into focused modules and align naming
- extract Chunker into chunking.rs alongside ChunkPipeline
- rename chunk_pipeline.rs to chunking.rs
- extract LiveTextVad and LiveTextSilence into live_text.rs
- rename long_dictation internal symbols to live_text throughout
- condense multi-line constant comments to one line each
