+++ src-tauri/src/audio/mod.rs
@@ -1,7 +1,8 @@
 pub mod audio;
-pub mod chunk_pipeline;
+pub mod chunking;
 pub mod clean_recording;
 pub mod helpers;
+pub mod live_text;
 pub mod microphone;
 pub mod pipeline;
 pub mod recorder;
@@ -10,5 +11,5 @@ pub mod streaming;
 pub mod types;
 
 pub use audio::*;
-pub use chunk_pipeline::*;
+pub use chunking::*;
 pub use pipeline::*;
